testing: move TestSectorPinningE2E
