+++ internal/testutils/cluster.go
@@ -1,4 +1,4 @@
-package test
+package testutils
 
 import (
 	"context"
