+++ internal/testutils/consensusnode.go
@@ -1,4 +1,4 @@
-package test
+package testutils
 
 import (
 	"testing"
