+++ internal/testutils/syncer.go
@@ -1,4 +1,4 @@
-package test
+package testutils
 
 import (
 	"context"
